move config_test
