+++ master/internal/config/config_test.go
@@ -1,4 +1,4 @@
-package internal
+package config
 
 import (
 	"bytes"
@@ -9,7 +9,6 @@ import (
 	"github.com/ghodss/yaml"
 	"gotest.tools/assert"
 
-	"github.com/determined-ai/determined/master/internal/config"
 	"github.com/determined-ai/determined/master/internal/db"
 	"github.com/determined-ai/determined/master/internal/provisioner"
 	"github.com/determined-ai/determined/master/internal/resourcemanagers"
@@ -44,7 +43,7 @@ resource_pools:
     task_container_defaults:
       dtrain_network_interface: if0
 `
-	expected := config.Config{
+	expected := Config{
 		Log: logger.Config{
 			Level: "info",
 			Color: false,
